// src/app/api/models/agent.ts
/* tslint:disable */
/* eslint-disable */
export interface Agent {
  aadhaar?: null | string;
  access?: boolean;
  approvalStatus?: null | string;
  assignedWorkflows?: null | string;
  bankId?: null | number;
  createdAt?: string;
  currentVersion?: null | string;
  email?: null | string;
  firstVersion?: null | string;
  fullName: string;
  id?: number;
  lastLogin?: null | string;
  lastUpdateDate?: null | string;
  mobile?: null | string;
  notes?: null | string;
  password?: null | string;
  photo?: null | string;
  role?: null | string;
  status?: null | string;
  updatedAt?: string;
}